Modified shoutOut to use use "message types".
Going forward messages logged with shoutOut should treat the second
argument as an indicator of message type, rather than as simply the
foreground color to use on the command line.

For backwards compatibility, the MsgType argument has the alias
"ForegroundColor", and can still be used as such: shoutOut will detect
when a valid color is used and apply it to the output using Write-Host.

Usage of "ForegroundCOlor" should be considered deprecated.