1st draft for a version with custom handlers
The idea is to make shoutOut extensible by giving it the ability to pass
packaged log messages to custom scriptblocks that may in turn call on
other logic to record and/or distribute message.

This should promote extensibility, and would also  allow us to offload
the calling thread by performing logging tasks (e.g. writing to file) in
a separate Thread.

This commit is made prior to implementing logFile/logHandler validation
with the _ensureShoutOutLogFile and _ensureShoutOutLogHandler functions.