Added ParameterSets, also fixed error reporting
Added the option to parse a string instead of a file, using the new
"String" ParameterSet. The old functionality is now the "File"
ParameterSet, and will be used by default unless the Content parameter
is specified.

Also fixed an error where the catch clause around the "include" case
would sometimes report the wrong include path.