Moved include files to the rendering step.
Files included using <<()>> will now be evaluated at rendering time, and
will be rendered as templates using the $Values from the including
scope.

This also means that these files will be added to the render cache.

The content generated when evauating .ps1 files will be treated as a
scripts to be executed at render-time.